Periscope2 Shiny Application Framework

Periscope2 is the next-generation package following the paradigm of the 'periscope' package to support a UI-standardized and rail-guarded enterprise quality application environment. This package also includes a variety of convenience functions for 'shiny' applications in a more modernized way. Base reusable functionality as well as UI paradigms are included to ensure a consistent user experience regardless of application or developer.

Details

'periscope2' differs from the 'periscope' package as follows:

Upgraded dependency on bootstrap v4 instead of bootstrap v3

New user modules (i.e. announcements)

More functionality and finer control over existing modules such as alert and reset

More control over customizing different application parts (header, footer, left sidebar, right sidebar and body)

Enhanced file structure to organize application UI, shiny modules, app configuration, .. etc

A gallery of 'periscope' and 'periscope2' example apps is hosted at http://periscopeapps.org

Function Overview

Create a new framework application instance:
create_application

Set application parameters in program/global.R:
set_app_parameters

Get any url parameters passed to the application:
get_url_parameters

Update an existing application with a needed sidebar:
create_left_sidebar
create_right_sidebar

Register user-created UI objects to the requisite application locations:
add_ui_body
add_ui_footer
add_ui_header
add_ui_left_sidebar
add_ui_right_sidebar

Included shiny modules with a customized UI:
downloadFileButton
downloadableTableUI
downloadablePlotUI
appResetButton
logViewerOutput

High-functionality standardized tooltips:
ui_tooltip

More Information

browseVignettes(package = 'periscope2')

See also

Author

Maintainer: Mohammed Ali mohammed@aggregate-genius.com

Other contributors:

Constance Brett [contributor]

Aggregate Genius Inc [sponsor]