Set Application Parameters

This function sets global parameters customizing the shiny application.

set_app_parameters(
  title,
  app_info = NULL,
  log_level = "DEBUG",
  app_version = "1.0.0",
  loading_indicator = NULL,
  announcements_file = NULL
)

Arguments

title

- Application title text

app_info

- Application detailed information. It can be character string, HTML value or NULL

A character string will be used to set a link target. This means the user will be able to click on the application title and be redirected in a new window to whatever value is given in the string. Any valid URL, File, or other script functionality that would normally be accepted in an <a href=...> tag is allowed.

An HTML value will be used to as the HTML content for a modal pop-up window that will appear on-top of the application when the user clicks on the application title.

Supplying NULL will disable the title link functionality.

log_level

- Designating the log level to use for the user log as 'DEBUG','INFO', 'WARN' or 'ERROR' (default = 'DEBUG')

app_version

- Character string designating the application version (default = '1.0.0')

loading_indicator

- It uses waiter (see https://waiter.john-coene.com/#/).
Pass a list like list(html = spin_1(), color = "#333e48") to
configure waiterShowOnLoad (refer to the package help for all styles).

announcements_file

- The path to announcements configuration file. Use announcementConfigurationsAddin to generate that file.

Value

no return value, called for setting new application global properties

Shiny Usage

Call this function from program/global.R to set the application parameters.

See also

Examples

  library(shiny)
  library(waiter)
  library(periscope2)

  # Inside program/global.R
  set_app_parameters(title              = "periscope Example Application",
                     app_info           = HTML("Example info"),
                     log_level          = "DEBUG",
                     app_version        = "1.0.0",
                     loading_indicator  = list(html = tagList(spin_1(), "Loading ...")),
                     announcements_file = "./program/config/announce.yaml")